+++ lib/weneedfeed/page_schema.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module Weneedfeed
+  PageSchema = Struct.new(
+    :id,
+    :item_description_selector,
+    :item_link_selector,
+    :item_time_selector,
+    :item_title_selector,
+    :item_selector,
+    :title,
+    :url,
+    keyword_init: true,
+  )
+end